requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
pyyaml>=6
joblib>=1.3
xgboost>=2.0
